rename app package
