+++ .gitignore
@@ -1,6 +1,7 @@
 .gradle
 /build/
 !gradle/wrapper/gradle-wrapper.jar
+out/
 
 ### STS ###
 .apt_generated
