+++ src/main/java/com/iobestgroup/donkeymoney/security/WebSecurityConfiguration.java
@@ -1,4 +1,4 @@
-package com.iobestgroup.aplikacjadozarzadzaniawydatkamiswoimiinietylko.security;
+package com.iobestgroup.donkeymoney.security;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
